refactor: Move root-password-rotation logic to own step and file
