+++ pkg/job/job.go
@@ -2,13 +2,33 @@ package job
 
 import (
 	"context"
+	"fmt"
 
 	mariadbv1alpha1 "github.com/mariadb-operator/mariadb-operator/v26/api/v1alpha1"
+	labels "github.com/mariadb-operator/mariadb-operator/v26/pkg/builder/labels"
 	"github.com/mariadb-operator/mariadb-operator/v26/pkg/metadata"
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
+
+func ListJobsForMariaDB(ctx context.Context, client ctrlclient.Client, mariadb *mariadbv1alpha1.MariaDB) (*batchv1.JobList, error) {
+	var jobList batchv1.JobList
+	if err := client.List(
+		ctx,
+		&jobList,
+		ctrlclient.InNamespace(mariadb.Namespace),
+		ctrlclient.MatchingLabels(
+			labels.NewLabelsBuilder().
+				WithMariaDBSelectorLabels(mariadb).
+				Build(),
+		),
+	); err != nil {
+		return nil, fmt.Errorf("error listing jobs with MariaDB (%s/%s): %v", mariadb.GetNamespace(), mariadb.GetName(), err)
+	}
+
+	return &jobList, nil
+}
 
 func ListJobs(ctx context.Context, client ctrlclient.Client,
 	backup *mariadbv1alpha1.PhysicalBackup) (*batchv1.JobList, error) {
@@ -50,6 +70,16 @@ func IsJobRunning(job *batchv1.Job) bool {
 		return false
 	}
 	return !IsJobFailed(job) && !IsJobSuspended(job) && !IsJobComplete(job)
+}
+
+func HasRunningJobs(jobList *batchv1.JobList) bool {
+	for _, job := range jobList.Items {
+		if job.Status.Active > 0 {
+			return true
+		}
+	}
+
+	return false
 }
 
 func IsStatusConditionTrue(conditions []batchv1.JobCondition, conditionType batchv1.JobConditionType) bool {
